Move fixtures to seperate module.
